pyproject.toml
[tool.poetry]
name = "library-app"
version = "0.1.0"
description = ""
authors = ["kirillovme <54503407+kirillovme@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
pytest = "^7.4.2"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
